// client/src/interface/room.ts
interface RoomData {
  id: string;
  first_fid: string;
  second_fid: string;
  sender_uid: string;
  displaymessage: string;
  updated_date: string;
  creation_date: string;
  visible: boolean;

  fid: string;
  uid: string;
  image: string;
  username: string;
  displayname: string;
}

export default RoomData;